+++ src/commands/openChange.ts
@@ -3,15 +3,18 @@ import { Resource } from "../resource";
 import IncomingChangeNode from "../treeView/nodes/incomingChangeNode";
 import { Command } from "./command";
 
-export class OpenChangeHead extends Command {
-  constructor() {
-    super("svn.openChangeHead");
+export class OpenChange extends Command {
+  constructor(
+    commandId: string,
+    private readonly against: string
+  ) {
+    super(commandId);
   }
 
   public async execute(
     arg?: Resource | Uri | IncomingChangeNode,
     ...resourceStates: SourceControlResourceState[]
   ) {
-    return this.openChange(arg, "HEAD", resourceStates);
+    return this.openChange(arg, this.against, resourceStates);
   }
 }